Preventing production db dropping to void

+++ copy_sitename_prod_to_beta.sh
@@ -106,12 +106,19 @@ fi
 # Export production database and import into beta database
 # ------------------------------------------------------------------------------
 
-echo "Goodie! Replacing database ${BETA_DATABASE_NAME}..."
+echo "Goodie! Now we Copy ${PROD_DATABASE_NAME} tables to ${BETA_DATABASE_NAME}..."
 
-mysql -e "drop database ${BETA_DATABASE_NAME}; create database ${BETA_DATABASE_NAME};"
-mysqldump ${PROD_DATABASE_NAME} | mysql ${BETA_DATABASE_NAME}
-
-echo "Great! Databases are in sync"
+if ("${PROD_DATABASE_NAME}" == "${BETA_DATABASE_NAME}"); then
+    echo Production: "${PROD_DATABASE_NAME}"
+    echo Beta: "${BETA_DATABASE_NAME}"
+    echo "Production database is same as Beta database, Check your vhost file, if enviroment variables has trailing whitespace."
+    echo "There is a risk of deleting your production database, if this script continue. Exit procedure.... "
+    exit 1
+else
+    mysql -e "drop database ${BETA_DATABASE_NAME}; create database ${BETA_DATABASE_NAME};"
+    mysqldump ${PROD_DATABASE_NAME} | mysql ${BETA_DATABASE_NAME}
+    echo "Great! Databases are in sync"
+fi
 
 
 # ------------------------------------------------------------------------------
